+++ srcs/parser/parser.c
@@ -6,7 +6,7 @@
 /*   By: tmatis <tmatis@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/10 23:03:16 by tmatis            #+#    #+#             */
-/*   Updated: 2021/05/23 10:35:40 by jmazoyer         ###   ########.fr       */
+/*   Updated: 2021/05/23 14:14:35 by tmatis           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -61,90 +61,27 @@ void	set_status_env(t_list **env_var, int status)
 }
 
 /*
-** Cree une liste de commande execute simultanement relier par des pipes
+** Geres les char "; | >> > <" : met un \33 (escape) devant
+** pour les differencier de l'user input
 */
 
-t_list	*pipes_commands(t_list *word_list, t_list *env_var)
+char	*special(char **str)
 {
-	t_list		*pipes_list;
-	int			special_id;
-	t_command	*command;
-	t_list		*elem;
+	char	*token;
 
-	pipes_list = NULL;
-	while (word_list)
-	{
-		special_id = escape_control(word_list->content);
-		if (special_id == PIPE)
-			word_list = word_list->next;
-		if (special_id == SEMICOLON || !word_list)
-			break ;
-		command = get_command(&word_list, env_var);
-		if (command)
-			elem = ft_lstnew(command);
-		if (!command || !elem)
-		{
-			ft_safe_free(command);
-			ft_lstclear(&pipes_list, free_command);
-			return (NULL);
-		}
-		ft_lstadd_back(&pipes_list, elem);
-	}
-	return (pipes_list);
-}
-
-t_list	*get_next_pipes(char **str, int *error, t_list **env_var)
-{
-	t_list	*token_list;
-	t_list	*pipes_list;
-
-	token_list = tokenize(str, error, *env_var, true);
-	if (!token_list)
-	{
-		ft_log_error(strerror(errno));
-		set_status_env(env_var, 2);
+	token = ft_calloc(4, sizeof (char));
+	if (!token)
 		return (NULL);
-	}
-	pipes_list = pipes_commands(token_list, *env_var);
-	ft_lstclear(&token_list, ft_safe_free);
-	if (!pipes_list)
+	token[0] = ESC;
+	if (is_special(*str) == APPEND)
 	{
-		ft_log_error(strerror(errno));
-		set_status_env(env_var, 2);
-		return (NULL);
+		ft_memcpy(token + 1, *str, 2);
+		(*str) += 2;
 	}
-	return (pipes_list);
-}
-
-/*
-** Fonction qui va call le tokensisateur et parse_commands et call l'exec
-*/
-
-int	exec_line(char *str, t_list **env_var)
-{
-	t_list		*word_list;
-	t_list		*pipes_list;
-	int			error;
-	int			return_value;
-
-	word_list = tokenize_all(str, &error, *env_var);
-	error_detector(word_list, &error);
-	ft_lstclear(&word_list, free);
-	if (error != NO_ERROR)
+	else
 	{
-		write_error(error);
-		set_status_env(env_var, 2);
-		return (0);
+		ft_memcpy(token + 1, *str, 1);
+		(*str) += 1;
 	}
-	while (*str)
-	{
-		pipes_list = get_next_pipes(&str, &error, env_var);
-		if (!pipes_list)
-			return (0);
-		return_value = exec(pipes_list, env_var);
-		ft_lstclear(&pipes_list, free_command);
-		if (return_value)
-			return (return_value);
-	}
-	return (0);
+	return (token);
 }
