+++ srcs/parser/parser.h
@@ -6,7 +6,7 @@
 /*   By: tmatis <tmatis@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/14 11:39:08 by tmatis            #+#    #+#             */
-/*   Updated: 2021/05/22 23:09:05 by jmazoyer         ###   ########.fr       */
+/*   Updated: 2021/05/23 14:08:38 by tmatis           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -63,4 +63,8 @@ t_list		*make_word(char **str, int *error, t_list *env_var,
 void		error_detector(t_list *tokens, int *error);
 void		free_table(char ***table);
 void		set_status_env(t_list **env_var, int status);
+t_bool		add_joint_tokens(t_list **dollar_tokens, t_list **tokens,
+									t_list **to_join);
+t_bool	append_dollar_token(t_list **dollar_tokens, t_list **tokens,
+								t_list **to_join, int append);
 #endif
